manus_work/requirements.txt
httpx
lxml
python-dateutil


